+++ parameter.h
@@ -8,20 +8,20 @@ class Parameter
 {
 public:
     Parameter();
-    Parameter(const std::string& name, const std::string& desc, const std::string& units, int64_t min, int64_t max, int64_t value, bool hex = false);
-    Parameter& operator=(int64_t d);
+    Parameter(const std::string& name, const std::string& desc, const std::string& units, double min, double max, double value, bool hex = false);
+    Parameter& operator=(double d);
     const std::string& name() const;
     const std::string& description() const;
     const std::string& unit() const;
     bool getIsHex() const;
-    int64_t getValue() const;
-    int64_t getRangeMin() const;
-    int64_t getRangeMax() const;
+    double getValue() const;
+    double getRangeMin() const;
+    double getRangeMax() const;
     size_t byteSize() const;
 private:
     std::string _name, _description, _unit;
     bool isHex;
-    int64_t _value, rangeMin, rangeMax;
+    double _value, rangeMin, rangeMax;
 };
 
 #endif // PARAMETER_H
